Backend-MS2/requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
sqlalchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.2.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
email-validator==2.1.1
python-dotenv==1.0.1
bcrypt==4.1.3
psycopg2-binary==2.9.10
pymysql==1.1.0
cryptography==42.0.7
